fix: correct Chip component typings (#1044)
As of version 2.15.2 the Chip component accepts the prop `selectedColor`
which it defines as an optional string prop. Though this change was
introduced to the Chip component itself, the typings for the component
were not updated which causes Typescript to complain when using the
`selectedColor` prop. This change adds the missing type to appease
Typescript.