fix: don't set statusbar style in Appbar.Header
The header shouldn't be setting statusbar style. It won't work properly when using libraries such as React Navigation since multiple screens can be mounted at one time, and user has no way to control it or integrate it with react Navigation's lifecycle this way.

If we really want to do this, a better way is to use the `<StatusBar />` component from React Native which will make sure that the style set gets cleaned up when the component unmounts. though it still won't work with React Navigation.

In addition, it's a side-effect and shouldn't be done in render.